Fix bad math in packed chunk pos
Still not sure how this happened or why unit test didn’t catch it

+++ src/main/java/grondag/exotic_matter/varia/PackedChunkPos.java
@@ -35,7 +35,7 @@ public class PackedChunkPos
 
     public static long getPackedChunkPosFromChunkXZ(int chunkX, int chunkZ)
     {
-            return (chunkX) + PackedChunkPos.CHUNK_BOUNDARY | ((chunkZ) + PackedChunkPos.CHUNK_BOUNDARY) << 32;
+            return (chunkX + PackedChunkPos.CHUNK_BOUNDARY) | (((long)chunkZ + PackedChunkPos.CHUNK_BOUNDARY) << 32);
     }
 
     public static ChunkPos unpackChunkPos(long packedChunkPos)
